+++ src/Service/ImageBuilder/src/ValueObject/ImageState.php
@@ -24,26 +24,42 @@ final class ImageState
     private $reason;
 
     /**
+     * The details about the failure, for images that failed to complete. Image Builder only sets this property when the
+     * image status is `FAILED`.
+     *
+     * @var ImageFailureContext|null
+     */
+    private $failureContext;
+
+    /**
      * @param array{
      *   status?: ImageStatus::*|null,
      *   reason?: string|null,
+     *   failureContext?: ImageFailureContext|array|null,
      * } $input
      */
     public function __construct(array $input)
     {
         $this->status = $input['status'] ?? null;
         $this->reason = $input['reason'] ?? null;
+        $this->failureContext = isset($input['failureContext']) ? ImageFailureContext::create($input['failureContext']) : null;
     }
 
     /**
      * @param array{
      *   status?: ImageStatus::*|null,
      *   reason?: string|null,
+     *   failureContext?: ImageFailureContext|array|null,
      * }|ImageState $input
      */
     public static function create($input): self
     {
         return $input instanceof self ? $input : new self($input);
+    }
+
+    public function getFailureContext(): ?ImageFailureContext
+    {
+        return $this->failureContext;
     }
 
     public function getReason(): ?string