+++ src/Service/ImageBuilder/src/ValueObject/ImageTestsConfiguration.php
@@ -9,8 +9,8 @@ namespace AsyncAws\ImageBuilder\ValueObject;
 final class ImageTestsConfiguration
 {
     /**
-     * Determines if tests should run after building the image. Image Builder defaults to enable tests to run following the
-     * image build, before image distribution.
+     * Specifies whether tests run after building the image. When enabled, tests run after the image build and before image
+     * distribution. Defaults to `true`.
      *
      * @var bool|null
      */